+++ src/main/java/com/connector/api/controller/profile/request/ProfileRequest.java
@@ -1,4 +1,4 @@
-package com.connector.api.domain.profile.request;
+package com.connector.api.controller.profile.request;
 
 import com.connector.api.domain.profile.ProfessionalStatus;
 import com.connector.api.domain.profile.Profile;
