+++ src/main/java/com/connector/api/service/profile/response/MyProfileResponse.java
@@ -1,4 +1,4 @@
-package com.connector.api.domain.profile.response;
+package com.connector.api.service.profile.response;
 
 import java.util.ArrayList;
 import java.util.List;
